Moved runes related stuff into their own module
Also, renamed id256 <--> btc transaction conversion methods to be more general
