+++ src/minter-did/src/id256.rs
@@ -7,9 +7,6 @@ use serde::Deserialize;
 
 use crate::error::Error;
 
-#[cfg(feature = "runes")]
-use ordinals::{RuneId};
-
 /// 32-bytes entity identifier.
 /// Uniquely identifies:
 /// - an EVM address,
@@ -46,7 +43,7 @@ impl Id256 {
     pub const BYTE_SIZE: usize = ID_256_BYTE_SIZE;
     pub const PRINCIPAL_MARK: u8 = 0;
     pub const EVM_ADDRESS_MARK: u8 = 1;
-    pub const RUNE_ID_MARK: u8 = 2;
+    pub const BTC_TX_MARK: u8 = 2;
 
     /// Creates unique identifier for contract.
     /// Chain id required to make identifiers unique across all chains.
@@ -105,30 +102,27 @@ impl Id256 {
         H160::from_slice(&bytes)
     }
 
-    /// Convert BTC rune id into id256 type.
+    /// Convert BTC transaction index into id256 type.
     ///
-    /// Runes are identified by block_id/transaction index in which the rune was etched.
-    ///
-    /// If `runes` feature is enabled, `rune_id.into()` can be used with the same effect.
-    pub fn from_rune_id(block_id: u64, tx_id: u32) -> Self {
+    /// Transactions in BTC network are indexed by the block id the transaction belongs to and the
+    /// index of the transaction in the block.
+    pub fn from_btc_tx_index(block_id: u64, tx_index: u32) -> Self {
         let mut buf = [0u8; Self::BYTE_SIZE];
 
-        buf[0] = Self::RUNE_ID_MARK;
+        buf[0] = Self::BTC_TX_MARK;
 
         let block_id_bytes = block_id.to_be_bytes();
-        let tx_id_bytes = tx_id.to_be_bytes();
+        let tx_id_bytes = tx_index.to_be_bytes();
         buf[1..][..block_id_bytes.len()].copy_from_slice(&block_id_bytes);
         buf[1 + block_id_bytes.len()..][..tx_id_bytes.len()].copy_from_slice(&tx_id_bytes);
 
         Self(buf)
     }
 
-    /// Converts Id256 into `(block_id, tx_id)` rune identifier if the ID represents the rune id,
+    /// Converts Id256 into `(block_id, tx_index)` transaction index if the ID represents the rune id,
     /// or returns an error otherwise.
-    ///
-    /// If `runes` feature is enabled, `id.try_into()` can be used with the same effect.
-    pub fn to_rune_id(&self) -> Result<(u64, u32), Error> {
-        if self.0[0] != Self::RUNE_ID_MARK {
+    pub fn to_btc_tx_index(&self) -> Result<(u64, u32), Error> {
+        if self.0[0] != Self::BTC_TX_MARK {
             return Err(Error::Internal("wrong rune id mark in Id256".into()));
         }
 
@@ -137,11 +131,11 @@ impl Id256 {
             .expect("we have exactly 8 bytes, as expected for u64");
         let block_id = u64::from_be_bytes(block_id_bytes);
 
-        let tx_id_bytes = self.0[9..13]
+        let tx_index_bytes = self.0[9..13]
             .try_into()
             .expect("we have exactly 8 bytes, as expected for u32");
-        let tx_id = u32::from_be_bytes(tx_id_bytes);
-        Ok((block_id, tx_id))
+        let tx_index = u32::from_be_bytes(tx_index_bytes);
+        Ok((block_id, tx_index))
     }
 }
 
@@ -218,23 +212,6 @@ impl Storable for Id256 {
     }
 }
 
-#[cfg(feature = "runes")]
-impl From<RuneId> for Id256 {
-    fn from(value: RuneId) -> Self {
-        Self::from_rune_id(value.block, value.tx)
-    }
-}
-
-#[cfg(feature = "runes")]
-impl TryFrom<Id256> for RuneId {
-    type Error = Error;
-
-    fn try_from(value: Id256) -> Result<Self, Self::Error> {
-        let (block, tx) = value.to_rune_id()?;
-        Ok(RuneId { block, tx })
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -307,26 +284,17 @@ mod tests {
     }
 
     #[test]
-    fn rune_id_id256_roundtrip() {
+    fn btc_tx_index_roundtrip() {
         let block_id = 100500;
         let tx_id = 42;
 
-        let id = Id256::from_rune_id(block_id, tx_id);
-        assert_eq!(id.to_rune_id(), Ok((block_id, tx_id)));
+        let id = Id256::from_btc_tx_index(block_id, tx_id);
+        assert_eq!(id.to_btc_tx_index(), Ok((block_id, tx_id)));
     }
 
     #[test]
-    fn rune_id_decode_error() {
+    fn btc_tx_index_decode_error() {
         let id = Id256::from_evm_address(&H160::from_slice(&[42; 20]), 31156);
-        assert!(matches!(id.to_rune_id(), Err(Error::Internal(_))), "unexpected to_rune_id result: {:?}", id.to_rune_id())
-    }
-
-    #[cfg(feature = "runes")]
-    #[test]
-    fn to_from_rune_id() {
-        let rune_id = RuneId { block: 256, tx: 42 };
-        let id = Id256::from(rune_id);
-
-        assert_eq!(id.try_into(), Ok(rune_id));
+        assert!(matches!(id.to_btc_tx_index(), Err(Error::Internal(_))), "unexpected to_rune_id result: {:?}", id.to_btc_tx_index())
     }
 }
