+++ src/minter-did/src/lib.rs
@@ -3,3 +3,7 @@ pub mod id256;
 pub mod init;
 pub mod order;
 pub mod reason;
+
+#[cfg(feature = "runes")]
+pub mod runes;
+
